🚚 api: move controller to certification/enrolment context
