+++ api/tests/certification/enrolment/unit/application/enrolment-controller_test.js
@@ -1,3 +1,5 @@
+import { sessionController } from '../../../../../lib/application/sessions/session-controller.js';
+import { usecases as libUsecases } from '../../../../../lib/domain/usecases/index.js';
 import { enrolmentController } from '../../../../../src/certification/enrolment/application/enrolment-controller.js';
 import { usecases } from '../../../../../src/certification/enrolment/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
@@ -55,6 +57,37 @@ describe('Certification | Enrolment | Unit | Controller | enrolment-controller',
         // then
         expect(response.statusCode).to.equal(201);
         expect(response.source).to.deep.equal(serializedCertificationCandidate);
+      });
+    });
+  });
+
+  describe('#importCertificationCandidatesFromCandidatesImportSheet', function () {
+    const sessionId = 2;
+    let request;
+    const odsBuffer = 'File Buffer';
+
+    beforeEach(function () {
+      // given
+      request = {
+        params: { id: sessionId },
+        payload: odsBuffer,
+      };
+
+      sinon.stub(libUsecases, 'importCertificationCandidatesFromCandidatesImportSheet').resolves();
+    });
+
+    it('should call the usecase to import certification candidates', async function () {
+      // given
+      libUsecases.importCertificationCandidatesFromCandidatesImportSheet.resolves();
+
+      // when
+      await sessionController.importCertificationCandidatesFromCandidatesImportSheet(request);
+
+      // then
+      expect(libUsecases.importCertificationCandidatesFromCandidatesImportSheet).to.have.been.calledWithExactly({
+        sessionId,
+        odsBuffer,
+        i18n: request.i18n,
       });
     });
   });
